+++ datasophon-service/src/main/java/com/datasophon/api/dto/extrepo/InstallComponentDTO.java
@@ -15,10 +15,10 @@ public class InstallComponentDTO {
 
     @NotNull(message = "meta文件不能为空")
     @Schema(description = "meta文件ID")
-    private Long meteFileId;
+    private Integer meteFileId;
 
     @Schema(description = "软件安装文件ID")
-    private Long pkgFileId;
+    private Integer pkgFileId;
 
     @NotBlank(message = "敏感文件解密密码不能为空")
     @Schema(description = "敏感文件解密密码")
